HARP-5947: Refactoring tile geometry creation * extraction into three new files * already has some structures that will make more sense   with phased loading and geoemtry toggling
