+++ @here/harp-mapview/index.ts
@@ -37,8 +37,8 @@ export * from "./lib/text/TextElementsRenderer";
 export * from "./lib/text/TextStyleCache";
 export * from "./lib/ThemeLoader";
 export * from "./lib/Tile";
-export * from "./lib/TileDataAccessor";
-export * from "./lib/TileGeometry";
+export * from "./lib/geometry/TileDataAccessor";
+export * from "./lib/geometry/TileGeometry";
 export * from "./lib/Utils";
 export * from "./lib/VisibleTileSet";
 export * from "./lib/WorkerBasedDecoder";
