+++ @here/harp-mapview/lib/MapView.ts
@@ -30,6 +30,7 @@ import { CopyrightInfo } from "./CopyrightInfo";
 import { DataSource } from "./DataSource";
 import { ElevationProvider } from "./ElevationProvider";
 import { ElevationRangeSource } from "./ElevationRangeSource";
+import { SimpleTileGeometryManager, TileGeometryManager } from "./geometry/TileGeometryManager";
 import { MapViewImageCache } from "./image/MapViewImageCache";
 import { MapViewFog } from "./MapViewFog";
 import { PickHandler, PickResult } from "./PickHandler";
@@ -555,6 +556,7 @@ export class MapView extends THREE.EventDispatcher {
     private m_elevationRangeSource?: ElevationRangeSource;
     private m_elevationProvider?: ElevationProvider;
     private m_visibleTileSetLock: boolean = false;
+    private m_tileGeometryManager: TileGeometryManager;
 
     private m_zoomLevel: number = DEFAULT_MIN_ZOOM_LEVEL;
     private m_minZoomLevel: number = DEFAULT_MIN_ZOOM_LEVEL;
@@ -803,7 +805,13 @@ export class MapView extends THREE.EventDispatcher {
             mapPassAntialiasSettings
         );
 
-        this.m_visibleTiles = new VisibleTileSet(this.m_camera, this.m_visibleTileSetOptions);
+        this.m_tileGeometryManager = new SimpleTileGeometryManager(this);
+
+        this.m_visibleTiles = new VisibleTileSet(
+            this.m_camera,
+            this.m_tileGeometryManager,
+            this.m_visibleTileSetOptions
+        );
 
         this.m_animatedExtrusionHandler = new AnimatedExtrusionHandler(this);
 
@@ -2456,7 +2464,11 @@ export class MapView extends THREE.EventDispatcher {
 
         this.calculateFocalLength(height);
 
-        this.m_visibleTiles = new VisibleTileSet(this.m_camera, this.m_visibleTileSetOptions);
+        this.m_visibleTiles = new VisibleTileSet(
+            this.m_camera,
+            this.m_tileGeometryManager,
+            this.m_visibleTileSetOptions
+        );
 
         // ### move & customize
         this.resize(width, height);
