+++ @here/harp-mapview/lib/geometry/TileGeometry.ts
