+++ @here/harp-omv-datasource/lib/OmvDebugLabelsTile.ts
@@ -3,7 +3,6 @@
  * Licensed under Apache 2.0, see full license in LICENSE
  * SPDX-License-Identifier: Apache-2.0
  */
-
 import { DecodedTile, getPropertyValue, isTextTechnique } from "@here/harp-datasource-protocol";
 import { TileKey } from "@here/harp-geoutils/lib/tiling/TileKey";
 import { DataSource, TextElement } from "@here/harp-mapview";
@@ -14,6 +13,8 @@ import {
     TextRenderStyle
 } from "@here/harp-text-canvas";
 import * as THREE from "three";
+
+import { TileGeometryCreator } from "../../harp-mapview/lib/geometry/TileGeometryCreator";
 import { OmvTile } from "./OmvTile";
 
 const debugMaterial = new THREE.LineBasicMaterial({
@@ -63,7 +64,9 @@ export class OmvDebugLabelsTile extends OmvTile {
      * @param decodedTile The decoded tile.
      */
     createTextElements(decodedTile: DecodedTile) {
-        super.createTextElements(decodedTile);
+        const geometryCreator = new TileGeometryCreator();
+
+        geometryCreator.createTextElements(this, decodedTile);
 
         const colorMap = new Map<number, THREE.Color>();
 
@@ -206,14 +209,5 @@ export class OmvDebugLabelsTile extends OmvTile {
                 }
             }
         }
-    }
-
-    /**
-     * Create the geometries for the decoded [[Tile]].
-     *
-     * @param decodedTile The decoded tile.
-     */
-    createGeometries(decodedTile: DecodedTile) {
-        super.createGeometries(decodedTile);
     }
 }
